uploading files to github
I made this first in repl.it and because it was finicky moved to my raspberry pi 3. This is not compatible with chromos because of its inability to use python idle (the official version of python for pc). Because of this, I recommend downloading the files to use on mac, windows or a linux platform that has python built in.